feat(backend): enhance agent scripts and improve configuration management
- Updated various agent scripts to improve functionality and maintainability, including enhancements to alert evaluation and notifications.
- Refactored configuration management for better handling of environment variables and integration with AWS services.
- Improved error handling and logging across multiple scripts to ensure better traceability and debugging.
- Added new tests to validate changes and ensure reliability of agent operations.

These changes aim to streamline agent operations and enhance overall system performance.

+++ backend/plane-b/src/repositories/implementations/quote-refresh-repository.ts
@@ -75,11 +75,11 @@ export class QuoteRefreshRepository implements IQuoteRefreshRepository {
           FOR UPDATE
        )
        UPDATE silver.quote_refresh_request AS req
-          SET status = $5,
-              locked_at = NOW(),
-              retry_count = CASE
-                WHEN $6 IS NULL THEN req.retry_count
-                ELSE GREATEST(req.retry_count, $6)
+        SET status = $5,
+            locked_at = NOW(),
+            retry_count = CASE
+                WHEN $6::int IS NULL THEN req.retry_count
+                ELSE GREATEST(req.retry_count, $6::int)
               END
         FROM next
         WHERE req.request_id = next.request_id
